+++ src/main/java/com/adnanumar/projects/airBnbApp/service/impl/RoomServiceImpl.java
@@ -3,7 +3,9 @@ package com.adnanumar.projects.airBnbApp.service.impl;
 import com.adnanumar.projects.airBnbApp.dto.RoomDto;
 import com.adnanumar.projects.airBnbApp.entity.Hotel;
 import com.adnanumar.projects.airBnbApp.entity.Room;
+import com.adnanumar.projects.airBnbApp.entity.User;
 import com.adnanumar.projects.airBnbApp.exception.ResourceNotFoundException;
+import com.adnanumar.projects.airBnbApp.exception.UnAuthorisedException;
 import com.adnanumar.projects.airBnbApp.repository.HotelRepository;
 import com.adnanumar.projects.airBnbApp.repository.RoomRepository;
 import com.adnanumar.projects.airBnbApp.service.InventoryService;
@@ -13,6 +15,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -39,6 +42,10 @@ public class RoomServiceImpl implements RoomService {
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " + hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This User does not own this Hotel with id : " + hotelId);
+        }
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -54,6 +61,10 @@ public class RoomServiceImpl implements RoomService {
         log.info("Getting all Rooms in hotel with id {}", hotelId);
         Hotel hotel = hotelRepository.findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with id : " + hotelId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This User does not own this Hotel with id : " + hotelId);
+        }
         return hotel.getRooms()
                 .stream()
                 .map((element) -> modelMapper.map(element, RoomDto.class))
@@ -75,6 +86,10 @@ public class RoomServiceImpl implements RoomService {
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with id : " + roomId));
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if (!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This User does not own this room with id : " + roomId);
+        }
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
         log.info("Room is deleted successfully with id {}", roomId);
